Fixes enemy collision
There were two problems, both relatively small. The collision shape for the hitbox needs to be bigger than the collision shape for the kinematic body if you want to be able to be hit before the physics engine takes over. Also, the name of the node is Enemy, not Enemy.tscn.
I noticed that your components in your player scene are creatively aligned (which may give you grief later), but I didn't fix that. Looking in Node2D->Transform->Position and making sure your point of reference in the scene is 0,0 will probably be helpful later on.